+++ src/main/webapp/wise5/directives/component/component.es6
@@ -79,7 +79,7 @@ class ComponentController {
             $scope.nodeAuthoringController = $scope.$parent.nodeAuthoringController;
         }
 
-        if ($scope.mode === 'authoring' && (componentContent.type == 'HTML' || componentContent.type == 'ConceptMap' || componentContent.type == 'Animation')) {
+        if ($scope.mode === 'authoring' && (componentContent.type == 'HTML' || componentContent.type == 'ConceptMap' || componentContent.type == 'Animation' || componentContent.type == 'Draw')) {
             $scope.componentTemplatePath = this.NodeService.getComponentAuthoringTemplatePath(componentContent.type);
         } else {
             $scope.componentTemplatePath = this.NodeService.getComponentTemplatePath(componentContent.type);
